decompose viewMore/addMore button
