--- src/components/Buttons/ViewMoreButton/ViewMoreButton.module.css
@@ -1,47 +1,0 @@
-.item {
-    width: 65px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    flex: 1 0 auto;
-    gap: 5px;
-}
-
-.item:hover {
-    background-size: 100%;
-    cursor: pointer;
-}
-.itemTitle {
-    composes: smallItemTitle from "@styles/style.css";
-
-}
-
-.item:hover>* {
-    color: var(--main-text);
-    border-color: var(--main-text);
-}
-
-.dashed {
-    margin-top: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    width: 43px;
-    height: 43px;
-    border: 3px dashed var(--text);
-    border-radius: 50%;
-    background-color: transparent;
-    color: var(--text);
-}
-
-.dashed i {
-    -webkit-text-stroke: 1.5px;
-}
-
-@media only screen and (max-width: 768px) {
-    .dashed {
-        width: 39px;
-        height: 39px;
-    }
-}
