+++ src/components/ModalWindows/ViewMoreModal/ViewMoreModal.module.css
@@ -35,55 +35,7 @@
     flex: 1 0 auto;
 }
 
-.dashed {
-    margin-top: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    width: 43px;
-    height: 43px;
-    border: 3px dashed var(--text);
-    border-radius: 50%;
-    background-color: transparent;
-    color: var(--text);
-}
-.item {
-    width: 65px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    flex: 1 0 auto;
-    gap: 5px;
-}
-
-.item:hover {
-    background-size: 100%;
-    cursor: pointer;
-}
-.itemTitle {
-    composes: smallItemTitle from "@styles/style.css";
-}
-.wrapped {
-    display: none;
-    visibility: hidden;
-    width: 0;
-    flex: 0 0;
-}
-
-.dashed i {
-    -webkit-text-stroke: 1.5px;
-}
-
-.specialItem .itemTitle {
-    margin-top: 4px;
-}
-
-.specialItem:hover>* {
-    color: var(--main-text);
-    border-color: var(--main-text);
-}
-
+.specialItem {}
 
 /* media */
 @media only screen and (max-width: 768px) {
@@ -91,10 +43,6 @@
         max-width: 510px;
         margin: 0 auto;
         padding: 0 15px;
-    }
-    .dashed {
-        width: 39px;
-        height: 39px;
     }
 
     .list {
