+++ src/pages/Dashboard/UserCategoriesCard/UserCategoriesCard.module.css
@@ -79,26 +79,8 @@
     flex: 1 0 auto;
 }
 
-.item {
-    width: 65px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    flex: 1 0 auto;
-    gap: 5px;
-}
-
-.item:hover {
-    background-size: 100%;
-    cursor: pointer;
-}
-
-.itemTitle {
-    composes: smallItemTitle from "@styles/style.css";
-}
-
 .wrapped {
-   display: none;
+   display: none !important;
    visibility: hidden;
    width: 0;
    flex: 0 0 0px;
