+++ src/pages/Dashboard/UserGroupsCard/UserGroupsCard.module.css
@@ -65,50 +65,17 @@
     flex-wrap: wrap;
 }
 
-.item {
-    width: 65px;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    flex: 1 0 auto;
-    gap: 5px;
+.specialItem h6{
+    order: -1;
+    margin-top: unset;
 }
-
-.item:hover {
-    background-size: 100%;
-    cursor: pointer;
-}
-
-.itemTitle {
-    composes: smallItemTitle from "@styles/style.css";
-}
-
-.specialItem:hover > * {
-    color: var(--main-text);
-    border-color: var(--main-text);
-}
-
-.dashed{
-    margin-bottom: auto;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    width: 43px;
-    height: 43px;
-    border: 3px dashed var(--text);
-    border-radius: 50%;
-    background-color: transparent;
-    color: var(--text);
-}
-
-.dashed i{
-    -webkit-text-stroke: 1.5px;
+.specialItem div {
+    margin-top: unset;
 }
 
 .wrapped {
-   display: none;
-}
+   display: none !important;
+} 
 
 .emptyList{
     display: flex;
@@ -134,10 +101,6 @@
 @media only screen and (max-width: 768px) {
     .inner{
         padding: 16px 16px 10px;
-    }
-    .dashed {
-        width: 39px;
-        height: 39px;
     }
     .list {
         gap: 8px;
