+++ src/pages/Dashboard/UserGroupsCardItem/UserGroupsCardItem.module.css
@@ -1,9 +1,20 @@
 .item {
-    composes: item from "@pages/Dashboard/UserGroupsCard/UserGroupsCard.module.css";
+    width: 65px;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    flex: 1 0 auto;
+    gap: 5px;
+
     background: linear-gradient(var(--item-hover) 0 0) no-repeat;
     background-size: 0%;
     border-radius: 7px;
 }
+.item:hover {
+    background-size: 100%;
+    cursor: pointer;
+}
+
 .itemInner {
     display: flex;
     flex-direction: column;
