+++ src/pages/Group/GroupCategoriesCard/GroupCategoriesCard.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useState } from 'react';
 //logic
 import { useElementSize } from 'usehooks-ts'
 import { json } from '@pages/Dashboard/UserCategoriesCard/objUserCategories';
@@ -8,6 +8,7 @@ import { handleWrap } from '@services/UsefulMethods/UIMethods';
 import classes from './GroupCategoriesCard.module.css';
 import CategoriesCardItem from '@components/CategoriesCardItem/CategoriesCardItem';
 import ExpenseModal from '@components/ModalWindows/ExpenseModal/ExpenseModal';
+import SpecialButton from '@components/Buttons/SpeciaButton/SpecialButton';
 
 
 
@@ -73,29 +74,25 @@ const GroupCategoriesCard = () => {
                     categories?.length === 0 ?
                         <div className={classes.emptyList}>
                             <p>Category list is empty!</p>
-                            <li className={`${classes.item} ${classes.specialItem}`}>
-                                <div className={classes.dashed}>
-                                    <i className="bi bi-plus-lg"></i>
-                                </div>
-                                <h6 className={classes.itemTitle}>Add More</h6>
-                            </li>
+                            <SpecialButton
+                                handleClick={() => {}}
+                                className={classes.specialItem}
+                                type='add'
+                            />
                         </div>
                         :
                         categories?.length! >= totalItems ?
-                        <li
-                            className={`${classes.item} ${classes.specialItem}`}>
-                            <div className={classes.dashed}>
-                                <i className="bi bi-chevron-right"></i>
-                            </div>
-                            <h6 className={classes.itemTitle}>View More</h6>
-                        </li>
+                         <SpecialButton 
+                            handleClick={() => {}}
+                            className={classes.specialItem}
+                            type='view'
+                        />
                         :
-                        <li className={`${classes.item} ${classes.specialItem}`}>
-                            <div className={classes.dashed}>
-                                <i className="bi bi-plus-lg"></i>
-                            </div>
-                            <h6 className={classes.itemTitle}>Add More</h6>
-                        </li>
+                        <SpecialButton
+                            handleClick={() => { }}
+                            className={classes.specialItem}
+                            type='add'
+                        />
                     }
                 </ul>
             </div>
